feat: Video Listing updated with Container and padding updates
